Move short ls helpers to the file manager.
This should resolve a small cyclic dependency.
